Fix moving inner blocks in the customizer
